URL and MySQL injection protection for deleting user in users.php page

+++ admin/includes/view_all_users.php
@@ -78,10 +78,11 @@
 if (isset($_GET['delete'])) 
 {
     
-
-        
-        
-            $delete_user_id = $_GET['delete'];
+    if(isset($_SESSION['user_role']))
+    {
+        if($_SESSION['user_role'] == 'Admin')
+        {
+            $delete_user_id = mysqli_real_escape_string($connection, trim($_GET['delete']));
 
             $query = "DELETE FROM users WHERE user_id = $delete_user_id LIMIT 1";
             $delete_result = mysqli_query($connection, $query);
@@ -89,6 +90,11 @@ if (isset($_GET['delete']))
             confirmQuery($delete_result);
 
             header("Location: users.php");
+        }
+    }
+        
+        
+            
         
 
         
@@ -99,26 +105,41 @@ if (isset($_GET['delete']))
 
 if (isset($_GET['change_to_admin'])) 
 {
-    $change_to_admin = $_GET['change_to_admin'];
+    if(isset($_SESSION['user_role']))
+    {
+        if($_SESSION['user_role'] == 'Admin')
+        {
+            $change_to_admin = mysqli_real_escape_string($connection, trim($_GET['change_to_admin']));
 
-    $query = "UPDATE users SET user_role = 'Admin' WHERE user_id = $change_to_admin LIMIT 1";
-    $change_to_admin_query = mysqli_query($connection, $query);
+            $query = "UPDATE users SET user_role = 'Admin' WHERE user_id = $change_to_admin LIMIT 1";
+            $change_to_admin_query = mysqli_query($connection, $query);
 
-    confirmQuery($change_to_admin_query);
+            confirmQuery($change_to_admin_query);
 
-    header("Location: users.php");
+            header("Location: users.php");
+        }
+    }
+
+    
 }
 
 if(isset($_GET['change_to_subscriber']))
 {
-    $change_to_subscriber = $_GET['change_to_subscriber'];
+    if(isset($_SESSION['user_role']))
+    {
+        if($_SESSION['user_role'] == 'Admin')
+        {
+            $change_to_subscriber = mysqli_real_escape_string($connection, trim($_GET['change_to_subscriber']));
 
-    $query = "UPDATE users SET user_role = 'Subscriber' WHERE user_id = $change_to_subscriber LIMIT 1";
-    $change_to_subscriber_query = mysqli_query($connection, $query);
+            $query = "UPDATE users SET user_role = 'Subscriber' WHERE user_id = $change_to_subscriber LIMIT 1";
+            $change_to_subscriber_query = mysqli_query($connection, $query);
 
-    confirmQuery($change_to_subscriber_query);
+            confirmQuery($change_to_subscriber_query);
 
-    header("Location: users.php");
+            header("Location: users.php");
+        }
+    }
+    
 }
 
 ?>
